[ style ] 정보 저장 함수 props로 내려주는 방법으로 수정
